requirements.txt
pyTelegramBotAPI
asyncio
aiohttp
xata
PyYAML
google-generativeai
